fix(ci): x86 size_t truncation, unused private field, portable death test
Three more issues caught by the CI matrix (not by the local
macOS build, since -Wconversion and -Wunused-private-field are
clang warnings):

1. include/gtlibcpp/freeze.hpp: FreezeStatus::original_value_u64
   was declared as std::size_t, which is 32-bit on x86. The
   bytes_to_u64() call returns a uint64_t and triggers MSVC C4244
   ("conversion from 'uint64_t' to 'size_t', possible loss of
   data") on every x86 build. Changed the field to std::uint64_t
   (it is the original *value* and the u64 suffix already says so).
   Updated the duplicate field in all_status() the same way.

2. src/named_pipe_transport.cpp: NamedPipeClient stored
   buffer_size_ but never read it. With -Wunused-private-field
   the build failed. Removed the field and the parameter name
   comment in the constructor.

3. tests/security_tests.cpp: the result_value_on_error_aborts
   test relied on std::signal(SIGABRT, ...) to convert the abort()
   call into a throwable, which works on POSIX but is not reliable
   on Windows (MSVC's abort() calls __fastfail / RaiseException,
   not raise(SIGABRT)). The test now:
   - always verifies that the failed Result reports the right
     error code and that value_or() returns the fallback
     (proving the failure is not masked),
   - on POSIX additionally installs a SIGABRT handler and
     confirms the abort path triggers,
   - on Windows the abort behaviour is verified by source
     inspection of result.hpp and the POSIX CI runs.

Local ctest: 4/4 suites, 42 tests, 100% pass.

+++ include/gtlibcpp/freeze.hpp
@@ -33,7 +33,7 @@ struct FreezeStatus {
     std::uint64_t value_u64{0};
     std::size_t  successful_rewrites{0};
     std::size_t  failed_rewrites{0};
-    std::size_t  original_value_u64{0};
+    std::uint64_t original_value_u64{0};
     bool         active{false};
     bool         restored{false};
     std::string  last_error{};
